Issues/batch5

+++ xCheckStudio/Webviewer/Javascripts/xCheckStudio/LogoutManager.js
@@ -1,7 +1,11 @@
 function onLogoutUser(alluser) {
 
-    return new Promise((resolve) => {
-        var userinfo = JSON.parse(localStorage.getItem('userinfo'));
+    return new Promise((resolve) => {      
+        let userinfo = xCheckStudio.Util.tryJsonParse(localStorage.getItem('userinfo'));
+        if (userinfo === null) {
+            return;
+        }
+        
         $.ajax({
             data: {
                 userid: userinfo.userid,
@@ -9,12 +13,14 @@ function onLogoutUser(alluser) {
             },
             type: "POST",
             url: "PHP/Logout.php",
-            success: function (msg) {
-                var object = JSON.parse(msg);
-                if (object.Msg === "Success") {
-                    if(object.AllUser === "No")
+            success: function (msg) {               
+                let object = xCheckStudio.Util.tryJsonParse(msg);
+
+                if (object !== null &&
+                    object.Msg === "Success") {
+                    if (object.AllUser === "No")
                         localStorage.clear();
-                    return resolve(true);                  
+                    return resolve(true);
                 }
                 return resolve(false);              
             },
